+++ src/components/Settings/Screen.js
@@ -1,0 +1,17 @@
+import React from "react";
+import { connect } from "react-redux";
+import Constants from "../../constants/index";
+
+class Screen extends React.Component {
+	render() {
+		return "Screen";
+	}
+}
+
+const mapStateToProps = (state) => {
+	return {
+		settings : state.settings
+	};
+};
+
+export default connect(mapStateToProps)(Screen);