+++ src/data/Settings.js
@@ -1,4 +1,16 @@
+import store from "../redux/store/index";
+import { updateState } from "../redux/actions/index";
+import localForage from "localforage";
+import merge from "lodash.merge";
+
+import ObjectService from "../services/ObjectService";
+
 const default_settings = {
+	intrinsic     : {
+		viewing    : "tables",
+		table_type : "dps",
+		collapsed  : false,
+	},
 	table_columns : {
 		dps  : [
 			"damage%",
@@ -87,17 +99,60 @@ const default_settings = {
 };
 
 class Settings {
-	constructor(settings) {
-		this.settings = settings || default_settings;
+	loadSettings() {
+		return new Promise((resolve, reject) => {
+			localForage.getItem("settings_cache")
+				.then((data) => {
+					let json                 = false;
+					let tmp_default_settings = Object.assign({}, default_settings);
+
+					try {
+						json = JSON.parse(data);
+					} catch (e) {}
+
+					if (!json) {
+						json = tmp_default_settings;
+					} else {
+						json = merge(tmp_default_settings, json);
+					}
+
+					this.settings = json;
+
+					store.dispatch(
+						updateState({
+							key   : "settings",
+							value : this.settings
+						})
+					);
+
+					resolve();
+				})
+				.catch((e) => {
+					reject();
+				});
+		});
+	}
+
+	saveSettings() {
+		return new Promise((resolve, reject) => {
+			localForage.setItem("settings_cache", JSON.stringify(this.settings || default_settings))
+				.then(() => {
+					resolve();
+				})
+				.catch((e) => {
+					reject();
+				});
+		});
 	}
 
 	getSetting(key) {
 		return this.settings[key] || default_settings[key];
 	}
 
-	setSetting(key, value) {
-		this.settings[key] = value;
+	setSetting(key_path, value) {
+		ObjectService.setByKeyPath(this.settings, key_path, value);
+		this.saveSettings();
 	}
 }
 
-export default Settings;
+export default new Settings();